pyproject.toml
[tool.poetry]
name = "course-5-db"
version = "0.1.0"
description = ""
authors = ["Max9222 <maxoog@bk.ru>"]
readme = "README.md"
packages = [{include = "course_5_db"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pip = "^23.2.1"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
